config/urls.py
from django.views.generic import TemplateView
from django.contrib import admin
from django.urls import path, include
from django.conf import settings


urlpatterns = [
    path("", TemplateView.as_view(template_name="index.html")),
    path("admin/", admin.site.urls),
    path("books/", include('apps.books.urls')),
]


if settings.DEBUG:
    import debug_toolbar

    urlpatterns += [
        path("__debug__/", include(debug_toolbar.urls)),
    ]
