requirements.txt
wheel==0.37.0
django==3.2.9
djangorestframework==3.12.4
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
whitenoise==5.3.0
